requirements.txt
# Core dependencies
numpy>=1.19.0
torch>=1.7.0
torchvision>=0.8.0
Pillow>=8.0.0
opencv-python>=4.5.0
scikit-learn>=0.24.0

# FastAPI dependencies
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0

# Utilities
python-dotenv>=0.19.0
aiofiles>=0.8.0 